+++ rapier3d/CHANGELOG.md
@@ -8,6 +8,13 @@ NOTE: This changelog is shared between bevy-tnua-rapier2d and bevy-tnua-rapier3d
 
 ## [Unreleased]
 
+### Removed
+- [**BREAKING**] `Default` implementation from `TnuaRapier2dPlugin` and
+  `TnuaRapier3dPlugin`. Since the schedule must match the physics backend
+  schedules, and since Avian and Rapier have different default schedules, it's
+  better for Tnua users to be actively aware which schedule they are operating
+  under.
+
 ## 0.12.0 - 2025-05-16
 ### Changed
 - Upgrade to Bevy 0.16 and bevy_rapier 0.30.